requirements.txt
Pillow>=8.0.0
shapely>=1.8.0
opencv-python
matplotlib
Flask>=2.0.0

